Extract TooltipView to its own file
